requirements.txt
# Core dependencies
google-cloud-bigquery==3.14.0
google-cloud-storage==2.13.0
google-cloud-pubsub==2.19.0
google-auth==2.25.0

# OCR and document processing
anthropic==0.20.0
pdf2image==1.17.0
Pillow==10.1.0

# API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Monitoring
prometheus-client==0.19.0

# Apache Airflow providers
apache-airflow-providers-google==10.10.0
apache-airflow-providers-dbt-cloud==3.5.0

# dbt
dbt-core==1.7.0
dbt-bigquery==1.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.13.0
mypy==1.7.1
pylint==3.0.3

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==23.2.0
